Assignment_2/requirements.txt
flask==2.0.1
werkzeug==2.0.1
mysql-connector-python==8.0.26
sentence-transformers==2.2.2
faiss-cpu==1.7.2
numpy==1.21.2
python-dotenv==0.19.0
transformers==4.30.2
torch==2.0.1
pytest==7.3.1
wikipedia==1.4.0
requests==2.31.0
# gradio==3.50.2
xformers